Random Button
Added a random button to select and fill random items, gems, enchants and modifiers.
Moved the validate code to statsStore for global use using the validateEntry function. This function validates if the item is already in one of the item slots and is not the same sub type as another item.